#1 Finish Refactor CreateAccount
