+++ src/interfaces/IEmptyMessage.ts
@@ -1,0 +1,7 @@
+export interface IEmptyMessage {
+  firstName: string;
+  lastName: string;
+  streetName: string;
+  buildingNumber: string;
+  country: string;
+}
