+++ src/utils/buildDinamicsStates.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from 'react';
+
+const buildHandler =
+  (setState: Dispatch<SetStateAction<string>>) =>
+  (e: React.FormEvent<HTMLInputElement>) =>
+    setState(e.currentTarget.value);
+
+export default buildHandler;
